feat(wayland): add KWinFT and wlroots plugins
Adds plugins that uses the Wrapland library to interact with compositors
supporting the kwinft_output_management_unstable_v1 or
wlr_output_management_unstable_v1 protocols.

The plugins are instantiated in parallel with the KWayland plugin and whichever
plugin has success in retrieving a management global first is used while the
other ones are rejected for this execution.