docs: correct license file
In aecdc24046b5a8d63d9ea10d425cfd8281644cbf the license was changed to
"LGPL v2.1 or later" but the added COPYING.LIB file contained obviously by
accident the text of the GPL v2 license.

Correct this now by providing the text of LGPL v2.1 in the file. Also remove
the superfluous COPYING file with the same text of the GPL v2. If a library
is released as LGPL v2.1, it can also be used in the constraints of GPL v2.