refactor: set output geometry
This change straightens out the Output API in regards to position, size and
scale of an output.

The consumer is allowed to set position, mode and scale of an output what
determines the output geometry in logical space.

Additionally a consumer might set a replication source.

Backends will evaluate all these values on change request to form a sensible
message to the windowing system.

On the other side there is no guarantee that a backend receives values for
scale or replication source of an output from the windowing system so consumers
should only write but not read out these and instead use other forms of data
synchronizations. In the future this could be moved into a separate control
class with a backend for some side storage.

BREAKING CHANGE: The Output API changes.