ci: add image trigger
On commits to master branch rebuild the KWinFT base image what includes Disman.
We ignore commits to one of the stable branches assuming there won't be many.

If we really need a rebuild of the image we can always rerun the last pipeline
for it manually in the ci-images repository.