feat: enable out-of-process backend mode by default
In the past the out-of-process backend mode was only enabled by default with
the X11 backend, but it works just as well with the KWinFT backend. Other
backends were not tested and in our autotests we must use the in-process mode.

This is a preliminary step for our remodel of the auxiliary properties flow.