feat: remove output cloning
Output cloning was only done on X11 but nowadays we use output replication on
Wayland and X11. On X11 this is a completely different code path.

It is unclear if and how clones were being made use of in the past on X11 but
since they seem to be not relevant anymore today remove the functionality to
reduce code complexity.