feat: redesign output modes API
New API functions are offered to query the last manual set and the "best" mode
which is determined by the maximum resolution with the highest refresh rate.

Additional combinations are possible by setting resolution or refresh rate
retrieval to automatic or manual.

For that we break with previous conventions and remove the need for specifying
a mode by id, instead we set modes by their defining attributes: resolution and
refresh rate.

BREAKING CHANGE: The Output API changes.