feat: establish control file write
With this large change we redesign the mode of operation of Disman and write
control files to a user directory when a configuration change was requested.

The reason is that not all properties of a certain output combination are
transferred through the window system endpoint, instead relying on some client-
side only property store.

When we do this in Disman directly we have the advantage of providing this
information to all its consumers simultaneously. This especially holds when
the Disman backend runs as a D-Bus service.

Closes #6