feat(api): add config generator
The Generator class provides methods to manipulate and optimize a provided
config in sensbile ways.

Backends make use of it to provide a fine preset configuration when no control
file is available. Frontend consumers can use it to improve upon the provided
backend configuration or change it up in other ways, for example on user input.