feat: add optional derived config when generating one
When a config is generated often it is with respect to a previous one that
should be respected in the sense that the settings of new outputs are best
possible generated while the properties of old outputs stay untouched.

Add functionality to the config class to either set all outputs in the config
we generate as derived ones or add - directly on creation - a separate config
specifying the old outputs.

Additionally for more control the utility functions for getting the primary and
biggest output receive optional exclusion lists of outputs that should not be
returned.