fix: prefer Wayland backend by WAYLAND_DISPLAY environment variable
The GUI application might be running in Xwayland mode, but checking
the WAYLAND_DISPLAY variable we can still select the Wayland backend
and communicate with the compositor directly.

Previously in this case the randr backend would have been selected what is
not the right one to talk to the compositor in a Wayland session.