feat: rename in-process environment variable
When a Disman-linking app is run it can be controlled if the app should try to
connect to Disman's backend via D-Bus and Disman's backend launcher process or
directly in-process create the backend.

This environement variable is renamed in this commit from
DISMAN_BACKEND_INPROCESS to DISMAN_IN_PROCESS.

The reason is that in-process is not a compound word so an underscore is in
order and that albeit that the varialbe name is more compact when the BACKEND
part is dropped.

BREAKING CHANGE: DISMAN_BACKEND_INPROCESS environement variable name changes to
DISMAN_IN_PROCESS.