fix: auto-scale to 110 DPI instead of 150 DPI base value
Feedback showed that the 150 DPI were unconvenient for normal desktop use.
Elements became too small on the screen.

The known standard historically is 96 DPI. But it feels this is a standard from
a different time and nowadays on HiDPI screens we can go with smaller items
without straining the eyes. So we try with 110 DPI what should increase the
size of elements drastically in comparison to 150 DPI but still give us more
screen space to use in comparison to sizing everything to 96 DPI.