fix: store and load output enabled value
The enabled value of an output must be saved and loaded from the control files.
We save it for now only per config and do not see it as a value that should be
globally saved per output. Though this might change in the future after
receiving some user feedback.