fix(randr): omit setting the output EDID
If set the EDID is used in Disman to generate the output hash from it. When it
is not set the name of the output is used as a fallback.

The EDID is currently making problems on control file read and write in the
RandR backend likely because it is not set all the time. We need to investigate
this further. For now disable setting it at all.