feat(wayland): set output description
Set in all available Wayland plugins the new description API field of the
Output class.

KWinFT and KWayland plugins need to set it to the output of our custom name
function. In the case of KWinFT a description event willl likely be added
to the protocol and the name field made clearer.

In the wlroots plugin for now we need to set the description as the name
because at the moment we fall back to using the name as an hashed identifier
of configs. Once we set the hash explicitly this can be corrected.