feat(lib): remove Edid handle in Output class
The Output class is lightened by the removal of its Edid component. All
relevant EDID data instead has been added to it directly.

This allows to remove lots of mechanism in our D-Bus API for retrieval of EDID
data per output. This slims the API down and makes it easier to maintain.