refactor(lib): access primary Output only by Config
Remove the getter and setter from Output class and instead handle all access
in the Config class.

This simplifies the code and we can say which output is primary is a property
of a config and not of a singular output. Additionally the concept of a primary
output does not exist anymore in Wayland what motivates to reduce the API
surface for it.