build: move Edid class to backend object library
The Edid class is only used in the backends and there specificially in the X11
plugin anymore. Move the class to the backend object library and remove it from
the public library interface. At the moment the Edid class can only be used
internal to the build and not in separate Wayland plugins.

BREAKING CHANGE: Edid class removed from public library.