feat: allow to set config file suffix
Add an std::string argument to the Filer ctor that defaults to an empty string
but if set to a non-empty string will be appended to the config file name that
is derived from the config hash.

This can be used in the future to write temporary configuration files mirroring
another one.