feat(lib): handle refresh rates in mHz
Instead of using a floating value to represent the refresh rate of an output
what was probably inspired by the QScreen API [1] move to using an integer
value of the refresh rate in mHz as is comnmon in Wayland protocols [2].

This simplifies the API and is more robust.

[1] https://doc.qt.io/qt-5/qscreen.html#refreshRate-prop
[2] https://wayland.freedesktop.org/docs/html/apa.html#protocol-spec-wl_output

BREAKING CHANGE: Refresh rates of modes are specified as integers in mHz.