fix: always emit config changed on frontend changed
A configuration change by a frontend client might only affect data that is not
relevant to the windowing system (for example the auto rotation value) but
still needs to be communicated to other clients.

As a quick fix for now in case the windowing system does not require changes
emit the config changed signal indiscriminately instead. This can be improved
in the future by checking if the new config has any data changed at all in
comparision to the current config and only then emit the signal.