fix: normalize positions on disabling embedded display
Wehn the embedded display is disabled the positions might be unequal to the
origin. To ensure compatibility with all windowing systems and compositors
shift the positions of remaining outputs relative to (0,0).

This is currently only done when disabling the embedded display but should be
valuable also in other contexts. For frontends this function should also be
offered via the API or wrapped by other more generic calls.