fix(lib): serialize resolution and refresh values by auto mode
The commanded mode might not return a valid mode when only the resolution or
refresh rate is set. But auto mode always returns a valid mode according to
set values and the auto settings.

So do it this way, what might lead to inconsistent data if the windowing system
did not accept our configuration request. But in that case tehre is nothing to
do for us anyway. Instead we have to provide correct configurations to the
windowing system and trust it to do the right thing afterwards.