style(wayland): check on set mode without local variable
Remove the success variable and just set the refresh rate in the
if-statement directly. The verbosity is not needed.

In the case of the removed assert in the wlroots backend this also fixes
a compiler warning in release builds.

(cherry picked from commit 7fd40ed4fd9969d6282898c2a95bfdbc9deb31b6)