fix(wayland): omit sending any other data on disabling wlr heads
The wlroots protocol is implicit about it but when a head is disabled by a
configuration request no other data should be sent for this head in the same
request.

The reason is that a configuration-head is then not well defined for which
the head would have been needed to be enabled first in this specific
configuration.

(cherry picked from commit 819f76866affb114f34dacfe707038813f7f2ae4)