fix(lib): prefer XDG_SESSION_TYPE for backend selection
If set prefer reading out the XDG_SESSION_TYPE environment variable to select
the backend to load.

This environment variable is likely to be updated when the windowing system is
being changed. This is not always the case with the WAYLAND_DISPLAY environment
variable.

(cherry picked from commit 7c7d95f8809fd4e347efe86e95a6b92635c2c314)