fix(lib): on backend selection check that Wayland socket exists
The WAYLAND_DISPLAY environment variable describes the name or the absolute
path of the socket Wayland clients should connect to.

When selecting the Wayland backend as preferred backend because of this
environment variable, validate this choice by ensuring the file for that socket
exists. If it does not exist the environment variable is likely out of date and
the Wayland backend can not be functional. In this case ignore it and continue
with X11.

(cherry picked from commit 1c3e15b7c4cfc3b2c119bae55f6f79fef755565d)