fix: only save primary value to config control file
This fixes the issue of the primary value being set both
in the config control file and in the output control file.
It should only save it to the config control file.
Not sure if this introduced any issues before, as
the this value was only read from the config control file.