fix: use std::string for EDID parsing
The previous QString lead to compile issues because of using a deleted QChar
constructor. Fix this by just using std::string with better API.

Since std::string is not as feature-rich we have to cleanup the string we
read in ourselves though. Do this through two simple loops removing trailing
whitespace and cutting the string off at a line break.