fix: always force geometry when replication is supported
When a backend supports output replication, the geometry of the replicating
output must be set to the one of the replication source.

This was done in the Wayland backend, but not the X11 backend. Instead of
reimplementing it there too, just lift the code up into BackendImpl's
set_config_impl function, so that it is run before sending the config further
to the backend.