build: rename wlroots plugin
While still making use of Wrapland we simply call it 'wlroots' as that's what
it is for and rename its directory.

The CMake target is renamed to 'disman-wlroots' in a push to unique target
names. We will change other targets later on to unique names too. The installed
plugin is simply called 'wlroots' again.