build: add toplevel custom target and install components
This simplifies the installation step by providing the 'disman' custom target
and several components to install. The custom target groups the targets of core
libraries and helpers and the components allows installation of this group of
targets as well as of additional plugin targets without knowing the exact
directories they are built into.