feat: support adaptive sync
We now provide adaptive sync support with the Wayland backend.

For that we add infrastructure to send a test commit on newly added outputs
that toggles adaptive sync and switches back again.

This gives us the information if the connected output supports adaptive sync
and afterwards we save this information in the emitted config to clients.